% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallow_conversation.R
\name{animal}
\alias{animal}
\title{Species Vocabulary}
\usage{
animal(species, sound)
}
\arguments{
\item{species}{Any animal species of your choice.}

\item{sound}{What sound does that animal make?}
}
\value{

}
\description{
Species Vocabulary
}
