% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallow_conversation.R
\name{food}
\alias{food}
\title{Baking Instructions}
\usage{
food(food, ingredient)
}
\arguments{
\item{food}{Enter a food item}

\item{ingredient}{Enter an ingredient for that food}
}
\value{

}
\description{
Baking Instructions
}
