% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallow_conversation.R
\name{shallow_conversation}
\alias{shallow_conversation}
\title{Tells you your name, shoe size, and birthday}
\usage{
shallow_conversation(name, shoe_size, birthday)
}
\arguments{
\item{name}{A character containing the input name}

\item{shoe_size}{A number containing one or two digits}

\item{birthday}{A date containing one chararcrter value (a month) and a number that contains one or 2 digits}
}
\value{

}
\description{
Contains one function that prints a statement about the user
}
