% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_uggo.R
\name{theme_uggo}
\alias{theme_uggo}
\title{Makes your graph ugly with poorly chosen colors}
\usage{
theme_uggo()
}
\value{

}
\description{
ugly text color, ugly line colors, and everything contrasts with everything
}
\examples{

}
